feat!: extract type-string parser into ext-php-rs-types crate
Slice 10 of issue #199. Splits the PHP type-hint parser out of the
runtime crate so the proc-macro crate can call it at expansion time and
parse-error reporting moves from extension load to `cargo build`.

New workspace member `crates/types` (package `ext-php-rs-types`) hosts
`PhpType`, `DnfTerm`, `PhpTypeParseError`, the recursive-descent parser,
`DataType`, and their `Display` and `Default` impls. The runtime crate
re-exports the moved items at their previous public paths
(`ext_php_rs::types::{PhpType, DnfTerm, PhpTypeParseError}` and
`ext_php_rs::flags::DataType`), so user code keeps compiling. Behind
a `proc-macro` cargo feature the new crate also exposes `quote::ToTokens`
impls that emit literal `PhpType::*` token trees referencing the runtime
crate's public paths; default features keep `quote`/`proc-macro2` out of
the runtime dep tree.

The macro `emit_phptype_from_str(lit: &LitStr) -> TokenStream`, which
emitted a runtime `from_str(LIT).unwrap_or_else(panic!)` call, is gone.
The replacement `parse_php_type_litstr(&LitStr) -> Result<PhpType>`
parses at expansion time and maps `PhpTypeParseError` to a `syn::Error`
spanned on the literal. The lightweight syntactic guard
`validate_php_types_litstr` plus its `PHP_TYPES_ALLOWED` and
`PHP_TYPES_MAX_LEN` constants are removed: the parser is the single
source of truth. `Function.returns_override` and
`TypedArg.php_type_override` storage flips from `Option<LitStr>` to
`Option<PhpType>`; emission becomes a plain `quote!(#parsed)` at the two
call sites in `function.rs` plus the matching paths in `impl_.rs` and
`interface.rs`.

BREAKING CHANGE: `DataType::as_u32`, `impl From<u32> for DataType`, and
`impl TryFrom<ZvalTypeFlags> for DataType` are removed. The orphan rule
forbids those impls from staying in the runtime crate now that
`DataType` lives in `ext-php-rs-types`. They are replaced by three free
functions in `ext_php_rs::flags`: `data_type_as_u32`,
`data_type_from_raw`, and `data_type_try_from_zvf`. Five internal call
sites are migrated; `dd-trace-php` does not use these conversions.

Workspace dependencies for `quote` and `proc-macro2` are hoisted to
`[workspace.dependencies]` since two crates now share them.

Closes the slice-06 acceptance gap: `#[php(types = "?Foo")]` and
similar parser-rejected strings now fail at `cargo build` with the
diagnostic spanned on the literal, not at first `cargo run`. The
"load-time panic" footnote in `guide/src/macros/function.md` is removed
and `guide/src/types/index.md` is updated to mention compile-time
parsing.
