+++ guide/src/macros/function.md
@@ -131,9 +131,9 @@ single Rust type via the `IntoZval`/`FromZval` trait path.
 
 The `#[php(types = "...")]` attribute on a parameter and the
 `#[php(returns = "...")]` attribute on a function override the registered PHP
-type metadata. The string is parsed at extension load by `PhpType::from_str`;
-the syntax matches the PHP type-hint grammar (with `\` for namespace
-separators).
+type metadata. The string is parsed at macro-expansion time by
+`PhpType::from_str`; the syntax matches the PHP type-hint grammar (with `\`
+for namespace separators).
 
 ```rust,ignore
 use ext_php_rs::prelude::*;
@@ -153,16 +153,11 @@ put `null` in the string if the parameter or return should be nullable. The
 runtime modifiers (`default`, `optional`, variadic, by-reference) are
 orthogonal to type and still apply.
 
-Validation runs in two stages:
-
-- **At macro-expansion time** the attribute is checked syntactically (LitStr
-  present, non-empty, allowed character set). Invalid input becomes a
-  `compile_error!` pointing at the literal.
-- **At extension load** the runtime parser builds the actual `PhpType`. A
-  parser-rejected string (for example `?Foo&Bar`, which the parser
-  refuses because class-side nullables aren't representable yet) panics with
-  the original literal in the message, surfacing on the first `cargo run` of
-  the consuming crate.
+Parsing runs once, at compile time. A parser-rejected string (for example
+`?Foo&Bar`, which the parser refuses because class-side nullables aren't
+representable as a single `PhpType`) becomes a `compile_error!` spanned on
+the literal — `cargo build` surfaces the diagnostic before the extension
+ever loads.
 
 The same attributes work inside `#[php_impl]`:
