+++ src/flags.rs
@@ -28,8 +28,6 @@ use crate::ffi::{
     ZEND_HAS_STATIC_IN_METHODS, ZEND_INTERNAL_FUNCTION, ZEND_USER_FUNCTION,
 };
 
-use std::{convert::TryFrom, fmt::Display};
-
 use crate::error::{Error, Result};
 
 bitflags! {
@@ -361,202 +359,131 @@ impl From<u8> for FunctionType {
     }
 }
 
-/// Valid data types for PHP.
-#[repr(C, u8)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Default)]
-pub enum DataType {
-    /// Undefined
-    Undef,
-    /// `null`
-    Null,
-    /// `false`
-    False,
-    /// `true`
-    True,
-    /// Integer (the irony)
-    Long,
-    /// Floating point number
-    Double,
-    /// String
-    String,
-    /// Array
-    Array,
-    /// Iterable
-    Iterable,
-    /// Object
-    Object(Option<&'static str>),
-    /// Resource
-    Resource,
-    /// Reference
-    Reference,
-    /// Callable
-    Callable,
-    /// Constant expression
-    ConstantExpression,
-    /// Void
-    #[default]
-    Void,
-    /// Mixed
-    Mixed,
-    /// Boolean
-    Bool,
-    /// Pointer
-    Ptr,
-    /// Indirect (internal)
-    Indirect,
-}
+pub use ext_php_rs_types::DataType;
 
-impl DataType {
-    /// Returns the integer representation of the data type.
-    #[must_use]
-    pub const fn as_u32(&self) -> u32 {
-        match self {
-            DataType::Undef => IS_UNDEF,
-            DataType::Null => IS_NULL,
-            DataType::False => IS_FALSE,
-            DataType::True => IS_TRUE,
-            DataType::Long => IS_LONG,
-            DataType::Double => IS_DOUBLE,
-            DataType::String => IS_STRING,
-            DataType::Array => IS_ARRAY,
-            DataType::Object(_) => IS_OBJECT,
-            DataType::Resource | DataType::Reference => IS_RESOURCE,
-            DataType::Indirect => IS_INDIRECT,
-            DataType::Callable => IS_CALLABLE,
-            DataType::ConstantExpression => IS_CONSTANT_AST,
-            DataType::Void => IS_VOID,
-            DataType::Mixed => IS_MIXED,
-            DataType::Bool => _IS_BOOL,
-            DataType::Ptr => IS_PTR,
-            DataType::Iterable => IS_ITERABLE,
-        }
+/// Returns the raw zval type-tag for `dt`.
+///
+/// Wraps the `IS_*` constants from `Zend/zend_types.h` so the rest of the
+/// crate calls a typed helper instead of poking at FFI integers. Lives as a
+/// free function in `crate::flags` rather than an inherent method on
+/// [`DataType`] because [`DataType`] now lives in the
+/// [`ext-php-rs-types`](https://crates.io/crates/ext-php-rs-types) workspace
+/// member, where the FFI constants are not in scope.
+#[must_use]
+pub const fn data_type_as_u32(dt: &DataType) -> u32 {
+    match dt {
+        DataType::Undef => IS_UNDEF,
+        DataType::Null => IS_NULL,
+        DataType::False => IS_FALSE,
+        DataType::True => IS_TRUE,
+        DataType::Long => IS_LONG,
+        DataType::Double => IS_DOUBLE,
+        DataType::String => IS_STRING,
+        DataType::Array => IS_ARRAY,
+        DataType::Object(_) => IS_OBJECT,
+        DataType::Resource | DataType::Reference => IS_RESOURCE,
+        DataType::Indirect => IS_INDIRECT,
+        DataType::Callable => IS_CALLABLE,
+        DataType::ConstantExpression => IS_CONSTANT_AST,
+        DataType::Void => IS_VOID,
+        DataType::Mixed => IS_MIXED,
+        DataType::Bool => _IS_BOOL,
+        DataType::Ptr => IS_PTR,
+        DataType::Iterable => IS_ITERABLE,
     }
 }
 
-// TODO: Ideally want something like this
-// pub struct Type {
-//     data_type: DataType,
-//     is_refcounted: bool,
-//     is_collectable: bool,
-//     is_immutable: bool,
-//     is_persistent: bool,
-// }
-//
-// impl From<u32> for Type { ... }
-
-impl TryFrom<ZvalTypeFlags> for DataType {
-    type Error = Error;
-
-    fn try_from(value: ZvalTypeFlags) -> Result<Self> {
-        macro_rules! contains {
-            ($t: ident) => {
-                if value.contains(ZvalTypeFlags::$t) {
-                    return Ok(DataType::$t);
-                }
-            };
-        }
-
-        contains!(Undef);
-        contains!(Null);
-        contains!(False);
-        contains!(True);
-        contains!(False);
-        contains!(Long);
-        contains!(Double);
-        contains!(String);
-        contains!(Array);
-        contains!(Resource);
-        contains!(Callable);
-        contains!(ConstantExpression);
-        contains!(Void);
-
-        if value.contains(ZvalTypeFlags::Object) {
-            return Ok(DataType::Object(None));
-        }
-
-        Err(Error::UnknownDatatype(0))
+/// Decodes a raw zval type-tag (`IS_*` constant) into a [`DataType`].
+///
+/// Replaces the previous `impl From<u32> for DataType`; orphan rules block
+/// that impl now that [`DataType`] lives in `ext-php-rs-types`.
+#[must_use]
+#[allow(clippy::bad_bit_mask)]
+pub fn data_type_from_raw(value: u32) -> DataType {
+    macro_rules! contains {
+        ($c: ident, $t: ident) => {
+            if (value & $c) == $c {
+                return DataType::$t;
+            }
+        };
     }
+
+    contains!(IS_VOID, Void);
+    contains!(IS_PTR, Ptr);
+    contains!(IS_INDIRECT, Indirect);
+    contains!(IS_CALLABLE, Callable);
+    contains!(IS_CONSTANT_AST, ConstantExpression);
+    contains!(IS_REFERENCE, Reference);
+    contains!(IS_RESOURCE, Resource);
+    contains!(IS_ARRAY, Array);
+    contains!(IS_STRING, String);
+    contains!(IS_DOUBLE, Double);
+    contains!(IS_LONG, Long);
+    contains!(IS_TRUE, True);
+    contains!(IS_FALSE, False);
+    contains!(IS_NULL, Null);
+
+    if (value & IS_OBJECT) == IS_OBJECT {
+        return DataType::Object(None);
+    }
+
+    contains!(IS_UNDEF, Undef);
+
+    DataType::Mixed
 }
 
-impl From<u32> for DataType {
-    #[allow(clippy::bad_bit_mask)]
-    fn from(value: u32) -> Self {
-        macro_rules! contains {
-            ($c: ident, $t: ident) => {
-                if (value & $c) == $c {
-                    return DataType::$t;
-                }
-            };
-        }
-
-        contains!(IS_VOID, Void);
-        contains!(IS_PTR, Ptr);
-        contains!(IS_INDIRECT, Indirect);
-        contains!(IS_CALLABLE, Callable);
-        contains!(IS_CONSTANT_AST, ConstantExpression);
-        contains!(IS_REFERENCE, Reference);
-        contains!(IS_RESOURCE, Resource);
-        contains!(IS_ARRAY, Array);
-        contains!(IS_STRING, String);
-        contains!(IS_DOUBLE, Double);
-        contains!(IS_LONG, Long);
-        contains!(IS_TRUE, True);
-        contains!(IS_FALSE, False);
-        contains!(IS_NULL, Null);
-
-        if (value & IS_OBJECT) == IS_OBJECT {
-            return DataType::Object(None);
-        }
-
-        contains!(IS_UNDEF, Undef);
-
-        DataType::Mixed
+/// Maps a [`ZvalTypeFlags`] bag onto the first matching [`DataType`].
+///
+/// Replaces the previous `impl TryFrom<ZvalTypeFlags> for DataType`.
+///
+/// # Errors
+///
+/// Returns [`Error::UnknownDatatype`] when no flag matches a known PHP type.
+pub fn data_type_try_from_zvf(value: ZvalTypeFlags) -> Result<DataType> {
+    macro_rules! contains {
+        ($t: ident) => {
+            if value.contains(ZvalTypeFlags::$t) {
+                return Ok(DataType::$t);
+            }
+        };
     }
-}
 
-impl Display for DataType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            DataType::Undef => write!(f, "Undefined"),
-            DataType::Null => write!(f, "Null"),
-            DataType::False => write!(f, "False"),
-            DataType::True => write!(f, "True"),
-            DataType::Long => write!(f, "Long"),
-            DataType::Double => write!(f, "Double"),
-            DataType::String => write!(f, "String"),
-            DataType::Array => write!(f, "Array"),
-            DataType::Object(obj) => write!(f, "{}", obj.as_deref().unwrap_or("Object")),
-            DataType::Resource => write!(f, "Resource"),
-            DataType::Reference => write!(f, "Reference"),
-            DataType::Callable => write!(f, "Callable"),
-            DataType::ConstantExpression => write!(f, "Constant Expression"),
-            DataType::Void => write!(f, "Void"),
-            DataType::Bool => write!(f, "Bool"),
-            DataType::Mixed => write!(f, "Mixed"),
-            DataType::Ptr => write!(f, "Pointer"),
-            DataType::Indirect => write!(f, "Indirect"),
-            DataType::Iterable => write!(f, "Iterable"),
-        }
+    contains!(Undef);
+    contains!(Null);
+    contains!(False);
+    contains!(True);
+    contains!(False);
+    contains!(Long);
+    contains!(Double);
+    contains!(String);
+    contains!(Array);
+    contains!(Resource);
+    contains!(Callable);
+    contains!(ConstantExpression);
+    contains!(Void);
+
+    if value.contains(ZvalTypeFlags::Object) {
+        return Ok(DataType::Object(None));
     }
+
+    Err(Error::UnknownDatatype(0))
 }
 
 #[cfg(test)]
 mod tests {
-    #![allow(clippy::unnecessary_fallible_conversions)]
-    use super::DataType;
+    use super::{DataType, data_type_from_raw};
     use crate::ffi::{
         IS_ARRAY, IS_ARRAY_EX, IS_CONSTANT_AST, IS_CONSTANT_AST_EX, IS_DOUBLE, IS_FALSE,
         IS_INDIRECT, IS_INTERNED_STRING_EX, IS_LONG, IS_NULL, IS_OBJECT, IS_OBJECT_EX, IS_PTR,
         IS_REFERENCE, IS_REFERENCE_EX, IS_RESOURCE, IS_RESOURCE_EX, IS_STRING, IS_STRING_EX,
         IS_TRUE, IS_UNDEF, IS_VOID,
     };
-    use std::convert::TryFrom;
 
     #[test]
     fn test_datatype() {
         macro_rules! test {
             ($c: ident, $t: ident) => {
-                assert_eq!(DataType::try_from($c), Ok(DataType::$t));
+                assert_eq!(data_type_from_raw($c), DataType::$t);
             };
         }
 
@@ -568,7 +495,7 @@ mod tests {
         test!(IS_DOUBLE, Double);
         test!(IS_STRING, String);
         test!(IS_ARRAY, Array);
-        assert_eq!(DataType::try_from(IS_OBJECT), Ok(DataType::Object(None)));
+        assert_eq!(data_type_from_raw(IS_OBJECT), DataType::Object(None));
         test!(IS_RESOURCE, Resource);
         test!(IS_REFERENCE, Reference);
         test!(IS_CONSTANT_AST, ConstantExpression);
@@ -579,7 +506,7 @@ mod tests {
         test!(IS_INTERNED_STRING_EX, String);
         test!(IS_STRING_EX, String);
         test!(IS_ARRAY_EX, Array);
-        assert_eq!(DataType::try_from(IS_OBJECT_EX), Ok(DataType::Object(None)));
+        assert_eq!(data_type_from_raw(IS_OBJECT_EX), DataType::Object(None));
         test!(IS_RESOURCE_EX, Resource);
         test!(IS_REFERENCE_EX, Reference);
         test!(IS_CONSTANT_AST_EX, ConstantExpression);
