+++ src/types/mod.rs
@@ -25,7 +25,7 @@ pub use iterator::ZendIterator;
 pub use long::ZendLong;
 pub use object::{PropertyQuery, ZendObject};
 pub use php_ref::PhpRef;
-pub use php_type::{DnfTerm, PhpType};
+pub use php_type::{DnfTerm, PhpType, PhpTypeParseError};
 pub use php_union::PhpUnion;
 pub use separated::Separated;
 pub use string::ZendStr;
